+++ src/electrifier/Controls/Vanara/ExplorerBrowser.xaml.cs
@@ -1,35 +1,25 @@
 /* todo: Use Visual States for Errors, Empty folders, Empty Drives */
 
-using System.Collections;
-using CommunityToolkit.Mvvm.Input;
-using CommunityToolkit.WinUI.Collections;
 using electrifier.Controls.Vanara.Services;
 using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Media.Imaging;
-using Microsoft.UI.Xaml;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
 using System.ComponentModel;
-using System.Data;
 using System.Diagnostics;
-using System.Drawing;
 using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices.WindowsRuntime;
 using System.Runtime.InteropServices;
-using System.Windows.Input;
-using electrifier.Controls.Vanara.Contracts;
+using electrifier.Controls.Vanara.Helpers;
 using Vanara.PInvoke;
 using Vanara.Windows.Shell;
-
 using static Vanara.PInvoke.Shell32;
 
 namespace electrifier.Controls.Vanara;
-using Visibility = Microsoft.UI.Xaml.Visibility;
+
 // https://github.com/dahall/Vanara/blob/master/Windows.Forms/Controls/ExplorerBrowser.cs
 /// <summary>Replacement for <see cref="Vanara.Windows.Forms.Controls.Explorer.cs">Windows.Forms/Controls/ExplorerBrowser.cs</see></summary>
 public sealed partial class ExplorerBrowser : INotifyPropertyChanged
 {
     private bool _isLoading;
+
     public bool IsLoading
     {
         get => _isLoading;
@@ -44,6 +34,7 @@ public sealed partial class ExplorerBrowser : INotifyPropertyChanged
             OnPropertyChanged();
         }
     }
+
     public event EventHandler<NavigatedEventArgs> Navigated;
     public event EventHandler<NavigationFailedEventArgs> NavigationFailed;
 
@@ -80,8 +71,8 @@ public sealed partial class ExplorerBrowser : INotifyPropertyChanged
                 var ebItem = new BrowserItem(child.PIDL, child.IsFolder)
                 {
                     SoftwareBitmap = child.IsFolder
-                    ? await GetStockIconBitmapSource(Shell32.SHSTOCKICONID.SIID_FOLDER)
-                    : await GetStockIconBitmapSource(Shell32.SHSTOCKICONID.SIID_DOCNOASSOC)
+                        ? await GetStockIconBitmapSource(Shell32.SHSTOCKICONID.SIID_FOLDER)
+                        : await GetStockIconBitmapSource(Shell32.SHSTOCKICONID.SIID_DOCNOASSOC)
                 };
 
                 target.ChildItems.Add(ebItem);
@@ -90,7 +81,8 @@ public sealed partial class ExplorerBrowser : INotifyPropertyChanged
         }
         catch (COMException comEx)
         {
-            Debug.Fail($"[Error] Navigate(<{target}>) failed. COMException: <HResult: {comEx.HResult}>: `{comEx.Message}`");
+            Debug.Fail(
+                $"[Error] Navigate(<{target}>) failed. COMException: <HResult: {comEx.HResult}>: `{comEx.Message}`");
             throw;
         }
         catch (Exception ex)
@@ -115,7 +107,8 @@ public sealed partial class ExplorerBrowser : INotifyPropertyChanged
 
             var siFlags = Shell32.SHGSI.SHGSI_LARGEICON | Shell32.SHGSI.SHGSI_ICON;
             var icninfo = Shell32.SHSTOCKICONINFO.Default;
-            SHGetStockIconInfo(shStockIconId, siFlags, ref icninfo).ThrowIfFailed($"SHGetStockIconInfo({shStockIconId})");
+            SHGetStockIconInfo(shStockIconId, siFlags, ref icninfo)
+                .ThrowIfFailed($"SHGetStockIconInfo({shStockIconId})");
 
             var hIcon = icninfo.hIcon;
             var icnHandle = hIcon.ToIcon();
@@ -151,7 +144,8 @@ public sealed partial class ExplorerBrowser : INotifyPropertyChanged
             Debug.Assert(addedItems.Count == 1);
             var selectedFolder = addedItems[0] as BrowserItem;
             var currentTreeNode = ShellTreeView.NativeTreeView.SelectedItem;
-            Debug.Print($".NativeTreeView_SelectionChanged(`{selectedFolder?.DisplayName}`, treeNode: {currentTreeNode?.ToString()}");
+            Debug.Print(
+                $".NativeTreeView_SelectionChanged(`{selectedFolder?.DisplayName}`, treeNode: {currentTreeNode?.ToString()}");
 
             if (currentTreeNode is BrowserItem browserItem && browserItem.PIDL.Equals(selectedFolder?.PIDL))
             {
@@ -173,12 +167,16 @@ public sealed partial class ExplorerBrowser : INotifyPropertyChanged
             }
         }
     }
+
     #region Property stuff
+
     private void OnPropertyChanged([CallerMemberName] string? propertyName = null)
     {
         PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
     }
+
     public event PropertyChangedEventHandler? PropertyChanged;
+
     private bool SetField<T>(ref T field, T value, [CallerMemberName] string? propertyName = null)
     {
         if (EqualityComparer<T>.Default.Equals(field, value))
@@ -190,57 +188,7 @@ public sealed partial class ExplorerBrowser : INotifyPropertyChanged
         OnPropertyChanged(propertyName);
         return true;
     }
+
     #endregion Property stuff
 }
 
-[DebuggerDisplay($"{{{nameof(ToString)}(),nq}}")]
-public class BrowserItem(Shell32.PIDL pidl, bool isFolder, List<AbstractBrowserItem<ShellItem>>? childItems = default)
-    : AbstractBrowserItem<ShellItem>(isFolder, childItems), INotifyPropertyChanged
-{
-    public readonly Shell32.PIDL PIDL = new(pidl);
-    public string DisplayName => ShellItem.GetDisplayName(ShellItemDisplayString.NormalDisplay) ?? ShellItem.ToString();
-    public ShellItem ShellItem = new(pidl);
-    public new ObservableCollection<BrowserItem> ChildItems = [];
-    public static BrowserItem FromPIDL(Shell32.PIDL pidl) => new(pidl, false);
-    public static BrowserItem FromShellFolder(ShellFolder shellFolder) => new(shellFolder.PIDL, true);
-    public static BrowserItem FromKnownFolderId(Shell32.KNOWNFOLDERID knownItemId) => new(new ShellFolder(knownItemId).PIDL, true);
-
-    public SoftwareBitmapSource? SoftwareBitmap;
-
-    public event PropertyChangedEventHandler? PropertyChanged;
-    protected virtual void OnPropertyChanged([CallerMemberName] string? propertyName = null)
-    {
-        PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
-    }
-    protected bool SetField<T>(ref T field, T value, [CallerMemberName] string? propertyName = null)
-    {
-        if (EqualityComparer<T>.Default.Equals(field, value)) return false;
-        field = value;
-        OnPropertyChanged(propertyName);
-        return true;
-    }
-}
-[DebuggerDisplay($"{{{nameof(ToString)}(),nq}}")]
-public partial class BrowserItemCollection : List<ShellItem>, IList
-{
-    protected IList ListImplementation => new List<BrowserItem>();
-
-    public void CopyTo(Array array, int index) => ListImplementation.CopyTo(array, index);
-    public new int Count => ListImplementation.Count;
-    public bool IsSynchronized => ListImplementation.IsSynchronized;
-    public object SyncRoot => ListImplementation.SyncRoot;
-    public int Add(object? value) => ListImplementation.Add(value);
-    public new void Clear() => ListImplementation.Clear();
-    public bool Contains(object? value) => ListImplementation.Contains(value);
-    public int IndexOf(object? value) => ListImplementation.IndexOf(value);
-    public void Insert(int index, object? value) => ListImplementation.Insert(index, value);
-    public void Remove(object? value) => ListImplementation.Remove(value);
-    public new void RemoveAt(int index) => ListImplementation.RemoveAt(index);
-    public bool IsFixedSize => ListImplementation.IsFixedSize;
-    public bool IsReadOnly => ListImplementation.IsReadOnly;
-    public new object? this[int index]
-    {
-        get => ListImplementation[index];
-        set => ListImplementation[index] = value;
-    }
-}
