+++ src/electrifier/Controls/Vanara/ShellNamespaceTreeControl.xaml
@@ -5,12 +5,13 @@
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:helpers="using:electrifier.Controls.Vanara.Helpers"
              d:DataContext="{d:DesignInstance Type=local:ExplorerBrowser}"
              mc:Ignorable="d">
     <Grid>
         <Grid.Resources>
             <DataTemplate x:Name="BrowserItemTemplate"
-                          x:DataType="local:BrowserItem">
+                          x:DataType="helpers:BrowserItem">
                 <TreeViewItem ItemsSource="{x:Bind ChildItems}">
                     <Grid>
                         <StackPanel HorizontalAlignment="Stretch"
